+++ CHANGES.md
@@ -6,6 +6,11 @@
 
 - `Black` now strips whitespace around `**` operators unless it's only operation in the expression (#2095)
 
+- `Black` now processes one-line docstrings by stripping leading and trailing spaces,
+  and adding a padding space when needed to break up """". (#1740)
+
+- `Black` now cleans up leading non-breaking spaces in comments (#2092)
+
 - `Black` now respects `--skip-string-normalization` when normalizing multiline
   docstring quotes (#1637)
 
@@ -46,6 +51,11 @@
 
 - PR #2053: Python 2 support is now optional, install with
   `python3 -m pip install black[python2]` to maintain support.
+
+- Exclude `venv` directory by default (#1683)
+
+- Fixed "Black produced code that is not equivalent to the source" when formatting
+  Python 2 docstrings (#2037)
 
 #### _Packaging_